Space相关功能

+++ src/main/java/com/dongdong/backend/entity/Blog.java
@@ -34,7 +34,7 @@ public class Blog {
     private String context;
 
     @Column(name="likes")
-    private Long likes;
+    private int likes;
 
     public Blog() {
